requirements.txt
Flask
pinecone-client==3.2.2
langchain-pinecone
langchain-google-genai
langchain-classic
langchain-core
langchain-community
langchain-text-splitters
langchain-huggingface
python-dotenv
pypdf
sentence-transformers
pyreadline3; sys_platform == 'win32'